+++ snakeAI/controller.h
@@ -2,7 +2,7 @@
 #define CONTROLLER_H
 #include <vector>
 #include "dqn.h"
-#include "policyGradient.h"
+#include "dpg.h"
 #include "ddpg.h"
 #include "ppo.h"
 #include "common.h"
@@ -12,7 +12,7 @@ class Controller
 {
 public:
     Controller(vector<vector<int> >& map);
-    void setState(vector<double>& statex,int x, int y, int xt, int yt);
+    void setState(vector<float>& statex,int x, int y, int xt, int yt);
     int AStarAgent(int x, int y, int xt, int yt);
     int randomSearchAgent(int x, int y, int xt, int yt);
     int dqnAgent(int x, int y, int xt, int yt);
@@ -20,13 +20,13 @@ public:
     int ddpgAgent(int x, int y, int xt, int yt);
     int ppoAgent(int x, int y, int xt, int yt);
     int bpAgent(int x, int y, int xt, int yt);
-    double reward1(int xi, int yi, int xn, int yn, int xt, int yt);
-    double reward2(int xi, int yi, int xn, int yn, int xt, int yt);
+    float reward1(int xi, int yi, int xn, int yn, int xt, int yt);
+    float reward2(int xi, int yi, int xn, int yn, int xt, int yt);
     bool move(int& x, int& y, int direct);
-    int maxAction(vector<double>& Action);
+    int maxAction(vector<float>& Action);
     vector<vector<int> >& map;
-    vector<double> state;
-    vector<double> nextState;
+    vector<float> state;
+    vector<float> nextState;
     DQN dqn;
     BPNet bp;
     DPG dpg;
